Term of Use
# privacy-policy
This application is used to inform visitors regarding our policies with the collection, use, and disclosure of Personal Information if anyone decided to use our Service. If you choose to use our Service, then you agree to the collection and use of information in relation to this policy. The Personal Information that we collect is used for providing and improving the Service. We will not use or share your information with anyone except as described in this Privacy Policy.
Information Collection and Use

Users' device information(Product model, operating system, advertising identifier(e.g. IDFA）
These informations are used to investigate bugs occuring on specific devices.
Advertising identifiers are used only by third party app providers to display adds.
IP addresses to detect malicious activities.
Security

We value your trust in providing us your Personal Information, thus we are striving to use commercially acceptable means of protecting it. But remember that no method of transmission over the internet, or method of electronic storage is 100% secure and reliable, and we cannot guarantee its absolute security.
Contact us

If you have any questions or suggestions about our Privacy Policy, do not hesitate to contact us at igpmanagertools@hotmail.com